pyproject.toml
[project]
name = "newfang-spine"
version = "0.1.0"
description = "Local-first AI project planning and documentation system."
authors = [
    { name = "Joshua Jenks" }
]
dependencies = [
    "typer[all]>=0.9.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "httpx>=0.24.0",
    "rich>=13.0.0",
    "python-multipart>=0.0.6",
    "pytest>=7.4.0",
]
requires-python = ">=3.9"

[project.scripts]
newfang = "newfang.cli.main:app"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build-meta"
